requirements.txt
Django>=3.2
djangorestframework>=3.12
python-multipart>=0.0.5
